requirements.txt
python-telegram-bot == 4.0.2
geopy == 1.11.0
nltk == 3.2
psycopg2 == 2.6.1
spacy == 0.100.6
